Rename core/documents.py to core/document.py
