+++ superduperdb/datalayer/mongodb/data_backend.py
@@ -2,7 +2,7 @@ import typing as t
 
 from pymongo import MongoClient
 
-from superduperdb.core.documents import Document
+from superduperdb.core.document import Document
 from superduperdb.core.serializable import Serializable
 from superduperdb.datalayer.base.data_backend import BaseDataBackend
 from superduperdb.misc.special_dicts import MongoStyleDict
