+++ tests/unittests/datalayer/mongodb/test_queries.py
@@ -2,7 +2,7 @@ import PIL.PngImagePlugin
 import pytest
 import torch
 
-from superduperdb.core.documents import Document
+from superduperdb.core.document import Document
 from superduperdb.datalayer.mongodb.query import Collection
 
 n_data_points = 250
